created ResetPowerLevelsCommand.java
